added some errors
