+++ client/src/components/Burger/OrderSummary/OrderSummary.tsx
@@ -1,7 +1,7 @@
 import React, {Component} from "react";
 import Auxiliary from '../../../hoc/Auxiliary/Auxiliary'
 import Button from '../../UI/Button/Button';
-import {Ingredient} from "../../../store/reducers/burgerBuilder";
+import {Ingredient} from "../../../store/actions/actionTypes";
 
 interface Props {
     ingredients: Ingredient,
