+++ client/src/containers/Checkout/Checkout.tsx
@@ -3,7 +3,7 @@ import {Redirect, Route, RouteComponentProps} from 'react-router-dom';
 import CheckoutSummary from '../../components/Order/CheckoutSummary/CheckoutSummary';
 import ContactData from "./ContactData/ContactData";
 import {connect} from 'react-redux';
-import {State} from "../../store/reducers/burgerBuilder";
+import {AppState} from "../../store/actions/actionTypes";
 
 interface ChildComponentProps extends RouteComponentProps<any> {
     ings: {}
@@ -37,7 +37,7 @@ class Checkout extends Component<ChildComponentProps> {
     }
 }
 
-const mapStateToProps = (state: State) => {
+const mapStateToProps = (state: AppState) => {
     return {
         ings: state.burgerBuilder.ingredients
     }
