+++ client/src/store/actions/actionTypes.ts
@@ -6,3 +6,21 @@ export const FETCH_INGREDIENTS_FAILED = 'FETCH_INGREDIENTS_FAILED';
 export const PURCHASE_BURGER_START = 'PURCHASE_BURGER_START';
 export const PURCHASE_BURGER_SUCCESS = 'PURCHASE_BURGER_SUCCESS';
 export const PURCHASE_BURGER_FAIL = 'PURCHASE_BURGER_FAIL';
+
+export type Ingredient = {
+    salad: number,
+    bacon: number,
+    cheese: number,
+    meat: number
+}
+
+export interface AppState {
+    burgerBuilder: {
+        ingredients: Ingredient | null,
+        totalPrice: number,
+        error: boolean
+    }
+    ingredients: Ingredient | null,
+    totalPrice: number,
+    error: boolean
+}
